+++ cachegrind/cg_sim.c
@@ -37,7 +37,7 @@
 */
 
 typedef struct {
-   Int          size;                   /* bytes */
+   Long          size;                   /* bytes */
    Int          assoc;
    Int          line_size;              /* bytes */
    Int          sets;
